refactor: QuizAnswerModal 내 AnswerCodeBlock 컴포넌트 분리
[#221]
