+++ packages/frontend/src/components/quiz/AnswerCodeBlock/index.ts
@@ -1,0 +1,1 @@
+export { AnswerCodeBlock } from "./AnswerCodeBlock";
