+++ packages/frontend/src/components/quiz/index.ts
@@ -3,3 +3,4 @@ export { default as QuizLocation } from "./QuizLocation";
 export { default as CommandAccordion } from "./CommandAccordion";
 export { SolvedModal, useSolvedModal } from "./SolvedModal";
 export { default as QuizAnswerModal } from "./QuizAnswerModal";
+export { AnswerCodeBlock } from "./AnswerCodeBlock";
